Issue #2163 - throw exception when filter cannot be applied / created
