+++ fhir-term-graph/src/test/java/com/ibm/fhir/term/graph/test/CodeSystemTermGraphLoaderTest.java
@@ -6,7 +6,6 @@
 
 package com.ibm.fhir.term.graph.test;
 
-import java.io.InputStream;
 import java.util.Collections;
 import java.util.LinkedHashSet;
 import java.util.Set;
@@ -15,8 +14,6 @@ import org.apache.commons.configuration.PropertiesConfiguration;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import com.ibm.fhir.model.format.Format;
-import com.ibm.fhir.model.parser.FHIRParser;
 import com.ibm.fhir.model.resource.CodeSystem;
 import com.ibm.fhir.model.resource.CodeSystem.Concept;
 import com.ibm.fhir.term.graph.FHIRTermGraph;
@@ -35,34 +32,29 @@ public class CodeSystemTermGraphLoaderTest {
     public void testCodeSystemTermGraphLoader() throws Exception {
         FHIRTermGraphUtil.setRootLoggerLevel(Level.INFO);
 
-        FHIRTermGraph graph = null;
-        try (InputStream in = CodeSystemTermGraphLoaderTest.class.getClassLoader().getResourceAsStream("JSON/CodeSystem-test.json")) {
-            graph = FHIRTermGraphFactory.open(new PropertiesConfiguration("conf/janusgraph-berkeleyje-lucene.properties"));
-            graph.dropAllVertices();
+        FHIRTermGraph graph = FHIRTermGraphFactory.open(new PropertiesConfiguration("conf/janusgraph-berkeleyje-lucene.properties"));
+        graph.dropAllVertices();
 
-            CodeSystem codeSystem = FHIRParser.parser(Format.JSON).parse(in);
-            FHIRTermGraphLoader loader = new CodeSystemTermGraphLoader(graph, codeSystem);
-            loader.load();
+        CodeSystem codeSystem = CodeSystemSupport.getCodeSystem("http://ibm.com/fhir/CodeSystem/test");
+        FHIRTermGraphLoader loader = new CodeSystemTermGraphLoader(graph, codeSystem);
+        loader.load();
 
-            FHIRTermServiceProvider provider = new GraphTermServiceProvider(graph);
+        FHIRTermServiceProvider provider = new GraphTermServiceProvider(graph);
 
-            Set<Concept> actual = new LinkedHashSet<>();
-            for (Concept concept : provider.getConcepts(codeSystem)) {
-                actual.add(provider.getConcept(codeSystem, concept.getCode()));
-            }
-
-            Set<Concept> expected = new LinkedHashSet<>();
-            for (Concept concept : CodeSystemSupport.getConcepts(codeSystem)) {
-                expected.add(concept.toBuilder()
-                    .concept(Collections.emptyList())
-                    .build());
-            }
-
-            Assert.assertEquals(actual, expected);
-        } finally {
-            if (graph != null) {
-                graph.close();
-            }
+        Set<Concept> actual = new LinkedHashSet<>();
+        for (Concept concept : provider.getConcepts(codeSystem)) {
+            actual.add(provider.getConcept(codeSystem, concept.getCode()));
         }
+
+        Set<Concept> expected = new LinkedHashSet<>();
+        for (Concept concept : CodeSystemSupport.getConcepts(codeSystem)) {
+            expected.add(concept.toBuilder()
+                .concept(Collections.emptyList())
+                .build());
+        }
+
+        Assert.assertEquals(actual, expected);
+
+        graph.close();
     }
 }
