+++ fhir-term-graph/src/test/java/com/ibm/fhir/term/graph/test/GraphTermServiceProviderTimeLimitTest.java
@@ -8,14 +8,10 @@ package com.ibm.fhir.term.graph.test;
 
 import static org.testng.Assert.fail;
 
-import java.io.InputStream;
-
 import org.apache.commons.configuration.PropertiesConfiguration;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-import com.ibm.fhir.model.format.Format;
-import com.ibm.fhir.model.parser.FHIRParser;
 import com.ibm.fhir.model.resource.CodeSystem;
 import com.ibm.fhir.term.graph.FHIRTermGraph;
 import com.ibm.fhir.term.graph.factory.FHIRTermGraphFactory;
@@ -25,6 +21,7 @@ import com.ibm.fhir.term.graph.provider.GraphTermServiceProvider;
 import com.ibm.fhir.term.graph.util.FHIRTermGraphUtil;
 import com.ibm.fhir.term.service.exception.FHIRTermServiceException;
 import com.ibm.fhir.term.spi.FHIRTermServiceProvider;
+import com.ibm.fhir.term.util.CodeSystemSupport;
 
 import ch.qos.logback.classic.Level;
 
@@ -34,11 +31,11 @@ public class GraphTermServiceProviderTimeLimitTest {
         FHIRTermGraphUtil.setRootLoggerLevel(Level.INFO);
 
         FHIRTermGraph graph = null;
-        try (InputStream in = GraphTermServiceProviderTimeLimitTest.class.getClassLoader().getResourceAsStream("JSON/CodeSystem-test.json")) {
+        try {
             graph = FHIRTermGraphFactory.open(new PropertiesConfiguration("conf/janusgraph-berkeleyje-lucene.properties"));
             graph.dropAllVertices();
 
-            CodeSystem codeSystem = FHIRParser.parser(Format.JSON).parse(in);
+            CodeSystem codeSystem = CodeSystemSupport.getCodeSystem("http://ibm.com/fhir/CodeSystem/test");
             FHIRTermGraphLoader loader = new CodeSystemTermGraphLoader(graph, codeSystem);
             loader.load();
